app/dynamic_node_id: avoid spurious reallocation
Only set a None unique ID in the DNA database if there's no existing
unique ID for the node ID and the node's unique ID is not available.
This prevents erasing an existing allocation and causing the node to be
allocated another ID, while still allowing the node ID to be reserved if
a unique ID never becomes available.